requirements.txt
# core packages
flask==1.1.2
flask_cors

# database
mongoengine==0.20.0

# API
marshmallow===3.3.0
flask_marshmallow===0.10.1
marshmallow-sqlalchemy===0.21.0
flask-apispec===0.8.8
apispec===3.3.0
webargs~=5.5.3

# deployment
gunicorn==20.0.4

# Error tracking
sentry-sdk[flask]

# cryptography
pycryptodome==3.9.8
